requirements.txt
streamlit
crewai
pyautogui
python-dotenv
langchain-google-genai
